backend/main.py
"""
VisionForge - 轻量级图像处理工具箱
FastAPI 后端服务
"""

from fastapi import FastAPI, UploadFile, File, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from PIL import Image
import io
import numpy as np

app = FastAPI(
    title="VisionForge API",
    description="轻量级图像处理工具箱 API",
    version="0.1.0"
)

# CORS 配置
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


@app.get("/")
async def root():
    """健康检查"""
    return {"status": "ok", "message": "VisionForge API is running"}


@app.get("/health")
async def health_check():
    """健康检查端点"""
    return {
        "status": "healthy",
        "version": "0.1.0",
        "features": ["image_enhancement", "face_processing", "object_detection"]
    }


# ==================== 图像增强接口 ====================

@app.post("/api/enhance/exposure")
async def adjust_exposure(image: UploadFile = File(...), factor: float = 1.0):
    """
    曝光调整
    - factor: 曝光因子 (0.5-2.0)
    """
    try:
        contents = await image.read()
        img = Image.open(io.BytesIO(contents))
        
        # 转换为 numpy 数组
        img_array = np.array(img, dtype=np.float32)
        
        # 调整曝光
        enhanced = img_array * factor
        enhanced = np.clip(enhanced, 0, 255).astype(np.uint8)
        
        # 返回结果
        output = Image.fromarray(enhanced)
        buf = io.BytesIO()
        output.save(buf, format='PNG')
        
        return {"status": "success", "message": "Exposure adjusted"}
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))


@app.post("/api/enhance/auto-correct")
async def auto_correct(image: UploadFile = File(...)):
    """
    自动色彩校正
    """
    try:
        contents = await image.read()
        img = Image.open(io.BytesIO(contents))
        
        # TODO: 实现自动色彩校正
        return {"status": "success", "message": "Auto correction applied"}
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))


# ==================== 人像处理接口 ====================

@app.post("/api/face/detect")
async def detect_faces(image: UploadFile = File(...)):
    """
    人脸检测
    """
    try:
        contents = await image.read()
        img = Image.open(io.BytesIO(contents))
        
        # TODO: 实现人脸检测
        return {
            "status": "success",
            "faces": [],
            "count": 0
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))


@app.post("/api/face/landmarks")
async def face_landmarks(image: UploadFile = File(...)):
    """
    人脸关键点检测
    """
    try:
        contents = await image.read()
        img = Image.open(io.BytesIO(contents))
        
        # TODO: 实现人脸关键点
        return {"status": "success", "landmarks": []}
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))


# ==================== 目标检测接口 ====================

@app.post("/api/detect/objects")
async def detect_objects(image: UploadFile = File(...)):
    """
    目标检测
    """
    try:
        contents = await image.read()
        img = Image.open(io.BytesIO(contents))
        
        # TODO: 实现目标检测
        return {
            "status": "success",
            "objects": [],
            "count": 0
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))


@app.post("/api/detect/segment")
async def segment_image(image: UploadFile = File(...)):
    """
    图像分割
    """
    try:
        contents = await image.read()
        img = Image.open(io.BytesIO(contents))
        
        # TODO: 实现图像分割
        return {"status": "success", "segments": []}
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))


if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
