backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pillow>=10.0.0
opencv-python>=4.8.0
onnxruntime>=1.15.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
